refact: move dto folder
