+++ src/main/java/floud/demo/dto/mypage/dto/MyFriend.java
@@ -1,12 +1,12 @@
-package floud.demo.dto.mypage;
+package floud.demo.dto.mypage.dto;
 
 import floud.demo.domain.enums.FriendshipStatus;
 import lombok.Builder;
 import lombok.Getter;
 
-@Getter
 @Builder
-public class MyWaiting {
+@Getter
+public class MyFriend {
     private Long friendship_id;
     private String nickname;
     private FriendshipStatus friendshipStatus;
